add all the tests
