+++ src/test/java/io/leonis/mx/MxWithValueNest1WithValueTest.java
@@ -2,12 +2,13 @@ package io.leonis.mx;
 
 import org.junit.*;
 
-public class MxNestWithoutValueTest {
+public class MxWithValueNest1WithValueTest {
+
   @Test
-  public void nest1mux6WithoutValue() throws Exception {
+  public void nest1mux6WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "1"))
             .add(marker -> marker + "2")
             .add(marker -> marker + "3")
@@ -16,15 +17,15 @@ public class MxNestWithoutValueTest {
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "@1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux1nest1mux5WithoutValue() throws Exception {
+  public void mux1nest1mux5WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "2"))
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
@@ -32,59 +33,59 @@ public class MxNestWithoutValueTest {
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1@2#3#4#5#6#7");
   }
 
   @Test
-  public void mux2nest1mux4WithoutValue() throws Exception {
+  public void mux2nest1mux4WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "3"))
             .add(marker -> marker + "4")
             .add(marker -> marker + "5")
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1#2@3#4#5#6#7");
   }
 
   @Test
-  public void mux3nest1mux3WithoutValue() throws Exception {
+  public void mux3nest1mux3WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
             .add(marker -> marker + "3")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "4"))
             .add(marker -> marker + "5")
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1#2#3@4#5#6#7");
   }
 
   @Test
-  public void mux4nest1mux2WithoutValue() throws Exception {
+  public void mux4nest1mux2WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "5"))
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1#2#3#4@5#6#7");
   }
 
   @Test
-  public void mux5nest1mux1WithoutValue() throws Exception {
+  public void mux5nest1mux1WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
@@ -92,15 +93,15 @@ public class MxNestWithoutValueTest {
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
             .add(marker -> marker + "5")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "6"))
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1#2#3#4#5@6#7");
   }
 
   @Test
-  public void mux6nest1WithoutValue() throws Exception {
+  public void mux6nest1WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
@@ -109,9 +110,9 @@ public class MxNestWithoutValueTest {
             .add(marker -> marker + "4")
             .add(marker -> marker + "5")
             .add(marker -> marker + "6")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "7"))
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1#2#3#4#5#6@7");
   }
 }
