+++ src/test/java/io/leonis/mx/MxWithValueNest1WithoutValueTest.java
@@ -2,14 +2,29 @@ package io.leonis.mx;
 
 import org.junit.*;
 
-public class MxNest2WithValueTest {
-
+public class MxWithValueNest1WithoutValueTest {
   @Test
-  public void nest2mux5WithValue() throws Exception {
+  public void nest1mux6WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
-            .join(Mx.mux("@")
-                .add(marker -> marker + "1")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "1"))
+            .add(marker -> marker + "2")
+            .add(marker -> marker + "3")
+            .add(marker -> marker + "4")
+            .add(marker -> marker + "5")
+            .add(marker -> marker + "6")
+            .add(marker -> marker + "7")
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
+        "#1#2#3#4#5#6#7");
+  }
+
+  @Test
+  public void mux1nest1mux5WithoutValue() throws Exception {
+    Assert.assertEquals(
+        Mx.mux("#")
+            .add(marker -> marker + "1")
+            .join(Mx.<String>mux()
                 .add(marker -> marker + "2"))
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
@@ -17,75 +32,59 @@ public class MxNest2WithValueTest {
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "@1@2#3#4#5#6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux1nest2mux4WithValue() throws Exception {
+  public void mux2nest1mux4WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
-            .join(Mx.mux("@")
-                .add(marker -> marker + "2")
+            .add(marker -> marker + "2")
+            .join(Mx.<String>mux()
                 .add(marker -> marker + "3"))
             .add(marker -> marker + "4")
             .add(marker -> marker + "5")
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1@2@3#4#5#6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux2nest2mux3WithValue() throws Exception {
+  public void mux3nest1mux3WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
-            .join(Mx.mux("@")
-                .add(marker -> marker + "3")
+            .add(marker -> marker + "3")
+            .join(Mx.<String>mux()
                 .add(marker -> marker + "4"))
             .add(marker -> marker + "5")
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2@3@4#5#6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux3nest2mux2WithValue() throws Exception {
-    Assert.assertEquals(
-        Mx.mux("#")
-            .add(marker -> marker + "1")
-            .add(marker -> marker + "2")
-            .add(marker -> marker + "3")
-            .join(Mx.mux("@")
-                .add(marker -> marker + "4")
-                .add(marker -> marker + "5"))
-            .add(marker -> marker + "6")
-            .add(marker -> marker + "7")
-            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3@4@5#6#7");
-  }
-
-  @Test
-  public void mux4nest2mux1WithValue() throws Exception {
+  public void mux4nest1mux2WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
-            .join(Mx.mux("@")
-                .add(marker -> marker + "5")
-                .add(marker -> marker + "6"))
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "5"))
+            .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4@5@6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux5nest2WithValue() throws Exception {
+  public void mux5nest1mux1WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
@@ -93,10 +92,26 @@ public class MxNest2WithValueTest {
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
             .add(marker -> marker + "5")
-            .join(Mx.mux("@")
-                .add(marker -> marker + "6")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "6"))
+            .add(marker -> marker + "7")
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
+        "#1#2#3#4#5#6#7");
+  }
+
+  @Test
+  public void mux6nest1WithoutValue() throws Exception {
+    Assert.assertEquals(
+        Mx.mux("#")
+            .add(marker -> marker + "1")
+            .add(marker -> marker + "2")
+            .add(marker -> marker + "3")
+            .add(marker -> marker + "4")
+            .add(marker -> marker + "5")
+            .add(marker -> marker + "6")
+            .join(Mx.<String>mux()
                 .add(marker -> marker + "7"))
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5@6@7");
+        "#1#2#3#4#5#6#7");
   }
 }
