+++ src/test/java/io/leonis/mx/MxWithValueNest2WithoutValueTest.java
@@ -2,30 +2,13 @@ package io.leonis.mx;
 
 import org.junit.*;
 
-public class MxNestWithValueTest {
-
+public class MxWithValueNest2WithoutValueTest {
   @Test
-  public void nest1mux6WithValue() throws Exception {
+  public void nest2mux5WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
-            .join(Mx.mux("@")
-                .add(marker -> marker + "1"))
-            .add(marker -> marker + "2")
-            .add(marker -> marker + "3")
-            .add(marker -> marker + "4")
-            .add(marker -> marker + "5")
-            .add(marker -> marker + "6")
-            .add(marker -> marker + "7")
-            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "@1#2#3#4#5#6#7");
-  }
-
-  @Test
-  public void mux1nest1mux5WithValue() throws Exception {
-    Assert.assertEquals(
-        Mx.mux("#")
-            .add(marker -> marker + "1")
-            .join(Mx.mux("@")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "1")
                 .add(marker -> marker + "2"))
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
@@ -33,75 +16,75 @@ public class MxNestWithValueTest {
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1@2#3#4#5#6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux2nest1mux4WithValue() throws Exception {
+  public void mux1nest2mux4WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
-            .add(marker -> marker + "2")
-            .join(Mx.mux("@")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "2")
                 .add(marker -> marker + "3"))
             .add(marker -> marker + "4")
             .add(marker -> marker + "5")
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2@3#4#5#6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux3nest1mux3WithValue() throws Exception {
+  public void mux2nest2mux3WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
-            .add(marker -> marker + "3")
-            .join(Mx.mux("@")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "3")
                 .add(marker -> marker + "4"))
             .add(marker -> marker + "5")
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3@4#5#6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux4nest1mux2WithValue() throws Exception {
+  public void mux3nest2mux2WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
             .add(marker -> marker + "3")
-            .add(marker -> marker + "4")
-            .join(Mx.mux("@")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "4")
                 .add(marker -> marker + "5"))
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4@5#6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux5nest1mux1WithValue() throws Exception {
+  public void mux4nest2mux1WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
-            .add(marker -> marker + "5")
-            .join(Mx.mux("@")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "5")
                 .add(marker -> marker + "6"))
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5@6#7");
+        "#1#2#3#4#5#6#7");
   }
 
   @Test
-  public void mux6nest1WithValue() throws Exception {
+  public void mux5nest2WithoutValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
@@ -109,10 +92,11 @@ public class MxNestWithValueTest {
             .add(marker -> marker + "3")
             .add(marker -> marker + "4")
             .add(marker -> marker + "5")
-            .add(marker -> marker + "6")
-            .join(Mx.mux("@")
+            .join(Mx.<String>mux()
+                .add(marker -> marker + "6")
                 .add(marker -> marker + "7"))
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6@7");
+        "#1#2#3#4#5#6#7");
   }
+
 }
