+++ src/test/java/io/leonis/mx/MxWithValueNest5WithValueTest.java
@@ -2,29 +2,14 @@ package io.leonis.mx;
 
 import org.junit.*;
 
-public class MxNest4WithoutValueTest {
-  @Test
-  public void nest4mux3WithoutValue() throws Exception {
-    Assert.assertEquals(
-        Mx.mux("#")
-            .join(Mx.<String>mux()
-                .add(marker -> marker + "1")
-                .add(marker -> marker + "2")
-                .add(marker -> marker + "3")
-                .add(marker -> marker + "4"))
-            .add(marker -> marker + "5")
-            .add(marker -> marker + "6")
-            .add(marker -> marker + "7")
-            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
-  }
+public class MxWithValueNest5WithValueTest {
 
   @Test
-  public void mux1nest4mux2WithoutValue() throws Exception {
+  public void nest5mux2WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
-            .add(marker -> marker + "1")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
+                .add(marker -> marker + "1")
                 .add(marker -> marker + "2")
                 .add(marker -> marker + "3")
                 .add(marker -> marker + "4")
@@ -32,38 +17,38 @@ public class MxNest4WithoutValueTest {
             .add(marker -> marker + "6")
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "@1@2@3@4@5#6#7");
   }
 
   @Test
-  public void mux2nest4mux1WithoutValue() throws Exception {
+  public void mux1nest5mux1WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
-            .add(marker -> marker + "2")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
+                .add(marker -> marker + "2")
                 .add(marker -> marker + "3")
                 .add(marker -> marker + "4")
                 .add(marker -> marker + "5")
                 .add(marker -> marker + "6"))
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1@2@3@4@5@6#7");
   }
 
   @Test
-  public void mux3nest4WithoutValue() throws Exception {
+  public void mux2nest5WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
             .add(marker -> marker + "2")
-            .add(marker -> marker + "3")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
+                .add(marker -> marker + "3")
                 .add(marker -> marker + "4")
                 .add(marker -> marker + "5")
                 .add(marker -> marker + "6")
                 .add(marker -> marker + "7"))
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1#2@3@4@5@6@7");
   }
 }
