+++ src/test/java/io/leonis/mx/MxWithValueNest6WithValueTest.java
@@ -2,12 +2,13 @@ package io.leonis.mx;
 
 import org.junit.*;
 
-public class MxNest6WithoutValueTest {
+public class MxWithValueNest6WithValueTest {
+
   @Test
-  public void nest6mux1WithoutValue() throws Exception {
+  public void nest6mux1WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "1")
                 .add(marker -> marker + "2")
                 .add(marker -> marker + "3")
@@ -16,15 +17,15 @@ public class MxNest6WithoutValueTest {
                 .add(marker -> marker + "6"))
             .add(marker -> marker + "7")
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "@1@2@3@4@5@6#7");
   }
 
   @Test
-  public void mux1nest6WithoutValue() throws Exception {
+  public void mux1nest6WithValue() throws Exception {
     Assert.assertEquals(
         Mx.mux("#")
             .add(marker -> marker + "1")
-            .join(Mx.<String>mux()
+            .join(Mx.mux("@")
                 .add(marker -> marker + "2")
                 .add(marker -> marker + "3")
                 .add(marker -> marker + "4")
@@ -32,22 +33,6 @@ public class MxNest6WithoutValueTest {
                 .add(marker -> marker + "6")
                 .add(marker -> marker + "7"))
             .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
-  }
-
-  @Test
-  public void mux2nest5WithoutValue() throws Exception {
-    Assert.assertEquals(
-        Mx.mux("#")
-            .add(marker -> marker + "1")
-            .add(marker -> marker + "2")
-            .join(Mx.<String>mux()
-                .add(marker -> marker + "3")
-                .add(marker -> marker + "4")
-                .add(marker -> marker + "5")
-                .add(marker -> marker + "6")
-                .add(marker -> marker + "7"))
-            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g),
-        "#1#2#3#4#5#6#7");
+        "#1@2@3@4@5@6@7");
   }
 }
