+++ src/test/java/io/leonis/mx/MxWithoutValueNest2WithValueTest.java
@@ -1,0 +1,108 @@
+package io.leonis.mx;
+
+import org.junit.*;
+
+public class MxWithoutValueNest2WithValueTest {
+
+  @Test
+  public void nest2mux5WithValue() throws Exception {
+    Assert.assertEquals(
+        Mx.<String>mux()
+            .expand(Mx.mux("@")
+                .expand(marker -> marker + "1")
+                .expand(marker -> marker + "2"))
+            .expand(marker -> marker + "3")
+            .expand(marker -> marker + "4")
+            .expand(marker -> marker + "5")
+            .expand(marker -> marker + "6")
+            .expand(marker -> marker + "7")
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g)
+            .apply("#"),
+        "@1@2#3#4#5#6#7");
+  }
+
+  @Test
+  public void mux1nest2mux4WithValue() throws Exception {
+    Assert.assertEquals(
+        Mx.<String>mux()
+            .expand(marker -> marker + "1")
+            .expand(Mx.mux("@")
+                .expand(marker -> marker + "2")
+                .expand(marker -> marker + "3"))
+            .expand(marker -> marker + "4")
+            .expand(marker -> marker + "5")
+            .expand(marker -> marker + "6")
+            .expand(marker -> marker + "7")
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g)
+            .apply("#"),
+        "#1@2@3#4#5#6#7");
+  }
+
+  @Test
+  public void mux2nest2mux3WithValue() throws Exception {
+    Assert.assertEquals(
+        Mx.<String>mux()
+            .expand(marker -> marker + "1")
+            .expand(marker -> marker + "2")
+            .expand(Mx.mux("@")
+                .expand(marker -> marker + "3")
+                .expand(marker -> marker + "4"))
+            .expand(marker -> marker + "5")
+            .expand(marker -> marker + "6")
+            .expand(marker -> marker + "7")
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g)
+            .apply("#"),
+        "#1#2@3@4#5#6#7");
+  }
+
+  @Test
+  public void mux3nest2mux2WithValue() throws Exception {
+    Assert.assertEquals(
+        Mx.<String>mux()
+            .expand(marker -> marker + "1")
+            .expand(marker -> marker + "2")
+            .expand(marker -> marker + "3")
+            .expand(Mx.mux("@")
+                .expand(marker -> marker + "4")
+                .expand(marker -> marker + "5"))
+            .expand(marker -> marker + "6")
+            .expand(marker -> marker + "7")
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g)
+            .apply("#"),
+        "#1#2#3@4@5#6#7");
+  }
+
+  @Test
+  public void mux4nest2mux1WithValue() throws Exception {
+    Assert.assertEquals(
+        Mx.<String>mux()
+            .expand(marker -> marker + "1")
+            .expand(marker -> marker + "2")
+            .expand(marker -> marker + "3")
+            .expand(marker -> marker + "4")
+            .expand(Mx.mux("@")
+                .expand(marker -> marker + "5")
+                .expand(marker -> marker + "6"))
+            .expand(marker -> marker + "7")
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g)
+            .apply("#"),
+        "#1#2#3#4@5@6#7");
+  }
+
+  @Test
+  public void mux5nest2WithValue() throws Exception {
+    Assert.assertEquals(
+        Mx.<String>mux()
+            .expand(marker -> marker + "1")
+            .expand(marker -> marker + "2")
+            .expand(marker -> marker + "3")
+            .expand(marker -> marker + "4")
+            .expand(marker -> marker + "5")
+            .expand(Mx.mux("@")
+                .expand(marker -> marker + "6")
+                .expand(marker -> marker + "7"))
+            .demux((a, b, c, d, e, f, g) -> a + b + c + d + e + f + g)
+            .apply("#"),
+        "#1#2#3#4#5@6@7");
+  }
+}
